refactor: move gui code to main file, create script file to demonstrate use without gui
